rename `mut_reference` to `unnecessary_mut_passed`
The lint was probably renamed at some point, but the files weren't. This
made it annoying to search for the lint.
