+++ clippy_lints/src/unnecessary_mut_passed.rs
