+++ tests/ui/unnecessary_mut_passed.fixed
