+++ tests/ui/unnecessary_mut_passed.rs
