+++ tests/ui/unnecessary_mut_passed.stderr
@@ -1,5 +1,5 @@
 error: the function `takes_ref` doesn't need a mutable reference
-  --> tests/ui/mut_reference.rs:56:15
+  --> tests/ui/unnecessary_mut_passed.rs:56:15
    |
 LL |     takes_ref(&mut 42);
    |               ^^^^^^^ help: remove this `mut`: `&42`
@@ -8,67 +8,67 @@ LL |     takes_ref(&mut 42);
    = help: to override `-D warnings` add `#[allow(clippy::unnecessary_mut_passed)]`
 
 error: the function `takes_ref_ref` doesn't need a mutable reference
-  --> tests/ui/mut_reference.rs:58:19
+  --> tests/ui/unnecessary_mut_passed.rs:58:19
    |
 LL |     takes_ref_ref(&mut &42);
    |                   ^^^^^^^^ help: remove this `mut`: `&&42`
 
 error: the function `takes_ref_refmut` doesn't need a mutable reference
-  --> tests/ui/mut_reference.rs:60:22
+  --> tests/ui/unnecessary_mut_passed.rs:60:22
    |
 LL |     takes_ref_refmut(&mut &mut 42);
    |                      ^^^^^^^^^^^^ help: remove this `mut`: `&&mut 42`
 
 error: the function `takes_raw_const` doesn't need a mutable reference
-  --> tests/ui/mut_reference.rs:62:21
+  --> tests/ui/unnecessary_mut_passed.rs:62:21
    |
 LL |     takes_raw_const(&mut 42);
    |                     ^^^^^^^ help: remove this `mut`: `&42`
 
 error: the function `as_ptr` doesn't need a mutable reference
-  --> tests/ui/mut_reference.rs:66:12
+  --> tests/ui/unnecessary_mut_passed.rs:66:12
    |
 LL |     as_ptr(&mut 42);
    |            ^^^^^^^ help: remove this `mut`: `&42`
 
 error: the function `as_ptr` doesn't need a mutable reference
-  --> tests/ui/mut_reference.rs:69:12
+  --> tests/ui/unnecessary_mut_passed.rs:69:12
    |
 LL |     as_ptr(&mut &42);
    |            ^^^^^^^^ help: remove this `mut`: `&&42`
 
 error: the function `as_ptr` doesn't need a mutable reference
-  --> tests/ui/mut_reference.rs:72:12
+  --> tests/ui/unnecessary_mut_passed.rs:72:12
    |
 LL |     as_ptr(&mut &mut 42);
    |            ^^^^^^^^^^^^ help: remove this `mut`: `&&mut 42`
 
 error: the function `as_ptr` doesn't need a mutable reference
-  --> tests/ui/mut_reference.rs:75:12
+  --> tests/ui/unnecessary_mut_passed.rs:75:12
    |
 LL |     as_ptr(&mut 42);
    |            ^^^^^^^ help: remove this `mut`: `&42`
 
 error: the method `takes_ref` doesn't need a mutable reference
-  --> tests/ui/mut_reference.rs:80:25
+  --> tests/ui/unnecessary_mut_passed.rs:80:25
    |
 LL |     my_struct.takes_ref(&mut 42);
    |                         ^^^^^^^ help: remove this `mut`: `&42`
 
 error: the method `takes_ref_ref` doesn't need a mutable reference
-  --> tests/ui/mut_reference.rs:82:29
+  --> tests/ui/unnecessary_mut_passed.rs:82:29
    |
 LL |     my_struct.takes_ref_ref(&mut &42);
    |                             ^^^^^^^^ help: remove this `mut`: `&&42`
 
 error: the method `takes_ref_refmut` doesn't need a mutable reference
-  --> tests/ui/mut_reference.rs:84:32
+  --> tests/ui/unnecessary_mut_passed.rs:84:32
    |
 LL |     my_struct.takes_ref_refmut(&mut &mut 42);
    |                                ^^^^^^^^^^^^ help: remove this `mut`: `&&mut 42`
 
 error: the method `takes_raw_const` doesn't need a mutable reference
-  --> tests/ui/mut_reference.rs:86:31
+  --> tests/ui/unnecessary_mut_passed.rs:86:31
    |
 LL |     my_struct.takes_raw_const(&mut 42);
    |                               ^^^^^^^ help: remove this `mut`: `&42`
